requirement.txt
fastapi
uvicorn
pillow
numpy
opencv-python
rembg
python-multipart
